fix: logging formatting (#292)
This pull request addresses a minor but important issue in logging by correcting a format specifier. The change ensures that the `offset` parameter in the `pwritev` function is correctly logged, preventing potential undefined behavior or incorrect output due to type mismatches with `printf`-style format strings. This improves the reliability and accuracy of trace logs.

### Highlights

* **Logging Format Fix**: Corrected the format specifier for the `offset` parameter in the `pwritev` function's trace logging to explicitly cast it to `intmax_t`, ensuring proper type matching for `%jd`.

(cherry picked from commit 63eee997375d8a2799412aae8458ceae177278a8)

+++ librawstorio/src/poll_queue.cpp
@@ -227,7 +227,8 @@ rawstor::io::Event* Queue::recvmsg(
     std::function<void(size_t, int)>&& cb
 ) {
     TraceEvent trace_event = RAWSTOR_TRACE_EVENT(
-        '|', "fd = %d, niov = %u, flags = %u\n", fd, msg->msg_iovlen, flags
+        '|', "fd = %d, niov = %u, flags = %u\n", fd,
+        (unsigned int)msg->msg_iovlen, flags
     );
     Session& s = _get_session(fd);
 
@@ -298,7 +299,7 @@ rawstor::io::Event* Queue::pwritev(
     std::function<void(size_t, int)>&& cb
 ) {
     TraceEvent trace_event = RAWSTOR_TRACE_EVENT(
-        '|', "fd = %d, niov = %u, offset = %jd\n", fd, niov, offset
+        '|', "fd = %d, niov = %u, offset = %jd\n", fd, niov, (intmax_t)offset
     );
     Session& s = _get_session(fd);
 
@@ -335,7 +336,8 @@ rawstor::io::Event* Queue::sendmsg(
     std::function<void(size_t, int)>&& cb
 ) {
     TraceEvent trace_event = RAWSTOR_TRACE_EVENT(
-        '|', "fd = %d, niov = %u, flags = %u\n", fd, msg->msg_iovlen, flags
+        '|', "fd = %d, niov = %u, flags = %u\n", fd,
+        (unsigned int)msg->msg_iovlen, flags
     );
     Session& s = _get_session(fd);
